refactor: InputView와 ResultView 클래스 view 패키지에 추가
