+++ src/main/kotlin/racingcar/README.md
@@ -25,5 +25,5 @@
 
 ## 5단계 기능 요구사항
 
-- [] InputView 와 ResultView View 패키지안에 넣기
-- [] MoveStrategy, VictoryCondition, Car, Setting Domain 패키지 안에 넣기
+- [x] InputView 와 ResultView View 패키지안에 넣기
+- [ ] MoveStrategy, VictoryCondition, Car, Setting Domain 패키지 안에 넣기
