+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -1,4 +1,6 @@
-package racingcar
+package racingcar.view
+
+import racingcar.Setting
 
 private const val MAXIMUM_WORD_COUNT = 5
