+++ src/test/kotlin/racingcar/view/InputViewTest.kt
@@ -1,8 +1,9 @@
-package racingcar
+package racingcar.view
 
 import io.kotest.assertions.throwables.shouldThrowWithMessage
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
+import racingcar.Setting
 
 class InputViewTest : BehaviorSpec({
